LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/reporting-to-document-reporting-endpoint.https.window-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/text-plain.txt?f01611bd-fa2b-4bfd-8932-c1cf37e5a6d1 because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/text-plain.txt?f01611bd-fa2b-4bfd-8932-c1cf37e5a6d1 due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html?3efd4143-1ea9-4979-84a8-7baaacb4e8ed because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html?4deb4a5c-fcd7-48a4-9fca-3b7d3749b53d' in a frame because of Cross-Origin-Embedder-Policy.

FAIL subresource CORP assert_true: A corp report with blockedURL https://127.0.0.1:9443/common/text-plain.txt and url https://localhost:9443/html/cross-origin-embedder-policy/resources/reporting-empty-frame.html?pipe=header(cross-origin-embedder-policy,require-corp;report-to=%22endpoint%22)|header(cross-origin-embedder-policy-report-only,require-corp;report-to=%22report-only-endpoint%22)|header(reporting-endpoints,%20endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=21cd8a9c-4da4-47d5-9af8-12c5e5599307%22\,%20report-only-endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=81bda1db-4864-4899-bcbe-dd62ee6746cf%22) is not found. expected true got false
FAIL navigation CORP on cross origin assert_true: A corp report with blockedURL https://127.0.0.1:9443/common/blank.html and url https://localhost:9443/html/cross-origin-embedder-policy/resources/reporting-empty-frame.html?pipe=header(cross-origin-embedder-policy,require-corp;report-to=%22endpoint%22)|header(cross-origin-embedder-policy-report-only,require-corp;report-to=%22report-only-endpoint%22)|header(reporting-endpoints,%20endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=21cd8a9c-4da4-47d5-9af8-12c5e5599307%22\,%20report-only-endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=81bda1db-4864-4899-bcbe-dd62ee6746cf%22) is not found. expected true got false
FAIL navigation CORP on same origin assert_true: A navigation report with blockedURL https://localhost:9443/common/blank.html and url https://localhost:9443/html/cross-origin-embedder-policy/resources/reporting-empty-frame.html?pipe=header(cross-origin-embedder-policy,require-corp;report-to=%22endpoint%22)|header(cross-origin-embedder-policy-report-only,require-corp;report-to=%22report-only-endpoint%22)|header(reporting-endpoints,%20endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=21cd8a9c-4da4-47d5-9af8-12c5e5599307%22\,%20report-only-endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=81bda1db-4864-4899-bcbe-dd62ee6746cf%22) is not found. expected true got false
